Update changes for Extended Union2
